Gemini Pandas Query Agent/requirements.txt
streamlit
pandas
python-dotenv
langchain_google_genai
langchain_experimental